+++ apps/server/src/controllers/rundownController.validate.ts
@@ -1,8 +1,6 @@
 import { body, param, validationResult } from 'express-validator';
-import { FastifySchema } from 'fastify';
-import { SupportedEvent } from 'ontime-types';
 
-const rundownPostValidator = [
+export const rundownPostValidator = [
   body('type').isString().exists().isIn(['event', 'delay', 'block']),
   (req, res, next) => {
     const errors = validationResult(req);
@@ -11,19 +9,7 @@ const rundownPostValidator = [
   },
 ];
 
-export const rundownPostSchema = {
-  body: {
-    type: 'object',
-    required: ['type'],
-    properties: {
-      type: { type: 'string', enum: [SupportedEvent.Block, SupportedEvent.Delay, SupportedEvent.Event] },
-      after: { type: 'string' },
-      id: { type: 'string' },
-    },
-  },
-} as const;
-
-const rundownPutValidator = [
+export const rundownPutValidator = [
   body('id').isString().exists(),
   (req, res, next) => {
     const errors = validationResult(req);
@@ -32,17 +18,7 @@ const rundownPutValidator = [
   },
 ];
 
-export const rundownPutSchema = {
-  body: {
-    type: 'object',
-    required: ['id'],
-    properties: {
-      id: { type: 'string' },
-    },
-  },
-} as const;
-
-const rundownReorderValidator = [
+export const rundownReorderValidator = [
   body('eventId').isString().exists(),
   body('from').isNumeric().exists(),
   body('to').isNumeric().exists(),
@@ -53,19 +29,7 @@ const rundownReorderValidator = [
   },
 ];
 
-export const rundownReorderSchema = {
-  body: {
-    type: 'object',
-    required: ['eventId', 'from', 'to'],
-    properties: {
-      eventId: { type: 'string' },
-      from: { type: 'integer', minimum: 0 },
-      to: { type: 'integer', minimum: 0 },
-    },
-  },
-} as const;
-
-const rundownSwapValidator = [
+export const rundownSwapValidator = [
   body('from').isString().exists(),
   body('to').isString().exists(),
   (req, res, next) => {
@@ -75,18 +39,7 @@ const rundownSwapValidator = [
   },
 ];
 
-export const rundownSwapSchema = {
-  body: {
-    type: 'object',
-    required: ['from', 'to'],
-    properties: {
-      from: { type: 'string' },
-      to: { type: 'string' },
-    },
-  },
-} as const;
-
-const paramsMustHaveEventId = [
+export const paramsMustHaveEventId = [
   param('eventId').exists(),
   (req, res, next) => {
     const errors = validationResult(req);
@@ -94,13 +47,3 @@ const paramsMustHaveEventId = [
     next();
   },
 ];
-
-export const paramsMustHaveEventIdSchema = {
-  params: {
-    type: 'object',
-    required: ['eventId'],
-    properties: {
-      eventId: { type: 'string' },
-    },
-  },
-} as const;
